pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "api-change-radar"
version = "0.1.0"
description = "A backend service that compares OpenAPI specs, detects changes, classifies risk, and produces structured reports."
readme = "README.md"
requires-python = ">=3.12"
authors = [
    { name = "Fatih Ayıbasan" }
]
dependencies = [
    "fastapi>=0.115,<1.0",
    "uvicorn[standard]>=0.30,<1.0",
    "pydantic>=2.8,<3.0",
    "pydantic-settings>=2.4,<3.0",
    "sqlalchemy>=2.0,<3.0",
    "alembic>=1.13,<2.0",
    "psycopg[binary]>=3.2,<4.0",
    "PyYAML>=6.0,<7.0",
    "openapi-spec-validator>=0.7,<1.0",
    "python-multipart>=0.0.9,<1.0",
    "structlog>=24.4,<25.0",
    "opentelemetry-api>=1.28,<2.0",
    "opentelemetry-sdk>=1.28,<2.0",
    "opentelemetry-exporter-otlp-proto-http>=1.28,<2.0",
    "opentelemetry-instrumentation-fastapi>=0.49b0,<1.0",
    "prometheus-client>=0.20,<1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3,<9.0",
    "pytest-cov>=5.0,<6.0",
    "httpx>=0.27,<1.0",
    "ruff>=0.6,<1.0",
]

[tool.setuptools.packages.find]
include = ["app*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
